Fixed preventing deletion of membershipOrganization
refs STA-307

+++ backend/app/api/src/endpoints/admin/organizations/PatchOrganizationsEndpoint.ts
@@ -51,6 +51,14 @@ export class PatchOrganizationsEndpoint extends Endpoint<Params, Query, Body, Re
                 throw new SimpleError({ code: 'not_found', message: 'Organization not found', statusCode: 404 });
             }
 
+            if (organization.id === (await Platform.getShared()).membershipOrganizationId) {
+                throw new SimpleError({
+                    code: 'cannot_delete_membership_organization',
+                    message: 'Cannot delete membership organization',
+                    human: 'Je kan de hoofdgroep niet verwijderen. Als je dit toch wil doen, kan je eerst een andere vereniging instellen als hoofdgroep via \'Boekhouding en aansluitingen\'.',
+                });
+            }
+
             await organization.delete();
         }
